% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_syntax.R
\name{coding_syntax}
\alias{coding_syntax}
\title{Outputs Coding Syntax}
\usage{
coding_syntax(data)
}
\arguments{
\item{data}{An Excel file with each tab containing the columns: id, variable name, code1, code2, code3. \emph{All columns should be present in each tab in the order shown.}}
}
\description{
Generates SPSS syntax for manually coded responses.
}
