% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_correlations.R
\name{reformat_correlations}
\alias{reformat_correlations}
\title{Reformat Correlation Matrix}
\usage{
reformat_correlations(input_file, cell_range, base_size)
}
\arguments{
\item{input_file}{Relative path of the correlation matrix.}

\item{cell_range}{Range of cells containing the relevant data. This is column-based and not row-based. \emph{Example: "A2:C35"}}

\item{base_size}{The total base size used to scale the correlations.}
}
\description{
Reformats the correlation matrix produced by SPSS into a more user-friendly format.
}
