% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_syntax.R
\name{spss_syntax}
\alias{spss_syntax}
\title{Outputs SPSS syntax}
\usage{
spss_syntax(data, method)
}
\arguments{
\item{data}{Dataset containing the variable names and labels.}

\item{method}{Type of SPSS syntax to generate. \emph{Valid methods: "var_labels" and "mr_sets"}}
}
\description{
Generates syntax for variable labels and MR Sets.
}
