% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_tables.R
\name{spss_tables}
\alias{spss_tables}
\title{Outputs SPSS tables syntax (Work in progress)}
\usage{
spss_tables(variable, type, title, banner, significance = TRUE)
}
\arguments{
\item{variable}{Name of the row variable.}

\item{type}{Type of custom table. Types currently supported are: \itemize{
\item Basic tables with values sorted in ascending order (Default)
\item Tables with T2B and B2B}}

\item{title}{Title of the table.}

\item{banner}{Character string specifying banner points seperated by commas.}

\item{significance}{When \code{significance = TRUE}, the default, signficance tables will be included.}
}
\description{
Generates syntax for custom tables in SPSS
}
\examples{
spss_tables("a1", "net", "Example title", "bannerPoint1, bannerPoint2, bannerPoint3", TRUE)
}
