% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straight_liners.R
\name{straight_liners}
\alias{straight_liners}
\title{Identifies straight liners in a survey.}
\usage{
straight_liners(data, qcode)
}
\arguments{
\item{data}{Dataset containing the responses of \strong{all} survey participants.}

\item{qcode}{Root question code including underscore. Multiple codes can be passed in as a character vector. \emph{Example: c("a2_", "a8_")}}
}
\description{
Identifies repondents who have straight lined through array-type questions.
}
