% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_speeders.R
\name{top_speeders}
\alias{top_speeders}
\title{Identifies speeders in a survey}
\usage{
top_speeders(data)
}
\arguments{
\item{data}{Dataset containing the responses of \strong{all} survey participants.}
}
\description{
Flags respondents who have completed the survey faster than 30% of the median.
}
